+++ tests/test_recipes_grating.py
@@ -65,6 +65,47 @@ def test_grating_2d_fit():
     assert np.isclose(cal.e0, true_offset, atol=1e-2, rtol=1e-2)
     assert np.isclose(cal.slope, true_slope, atol=1e-2, rtol=1e-2)
 
+def test_reading_grating2d(mock_sqs_grating_calibration_run, tmp_path):
+    monochromator_energy = "SA3_XTD10_MONO/MDL/PHOTON_ENERGY"
+    final_photon_spectrometer = "SQS_DIAG3_BIU/CAM/CAM_6:daqOutput"
+
+    monochromator_scan = Scan(mock_sqs_grating_calibration_run[monochromator_energy, "actualEnergy"], resolution=1)
+    grating_calibration = Grating2DCalibration(angle=0.0)
+    grating_calibration.setup(mock_sqs_grating_calibration_run[final_photon_spectrometer, "data.image.pixels"],
+                              monochromator_scan,
+                             )
+    d = tmp_path / "data"
+    d.mkdir()
+    fpath = str(d / "grating2d_test.h5")
+    grating_calibration.to_file(fpath)
+    grating_calibration = Grating2DCalibration.from_file(fpath)
+
+    calibrated = grating_calibration.apply(mock_sqs_grating_calibration_run(np.s_[10:20]))
+
+    assert np.isclose(grating_calibration.e0, 990.0, atol=1e-2, rtol=1e-2)
+    assert np.isclose(grating_calibration.slope, 20.0/1000.0, atol=1e-2, rtol=1e-2)
+
+def test_reading_grating1d(mock_sqs_grating_calibration_run, tmp_path):
+    monochromator_energy = "SA3_XTD10_MONO/MDL/PHOTON_ENERGY"
+    final_photon_spectrometer = "SQS_EXP_GH2-2/CORR/RECEIVER"
+
+    monochromator_scan = Scan(mock_sqs_grating_calibration_run[monochromator_energy, "actualEnergy"], resolution=1)
+    grating_calibration = Grating1DCalibration(min_pipxel=0, max_pixel=1000)
+    grating_calibration.setup(mock_sqs_grating_calibration_run[final_photon_spectrometer, "data.image.pixels"],
+                              monochromator_scan,
+                             )
+    d = tmp_path / "data"
+    d.mkdir()
+    fpath = str(d / "grating1d_test.h5")
+    grating_calibration.to_file(fpath)
+    grating_calibration = Grating1DCalibration.from_file(fpath)
+
+    calibrated = grating_calibration.apply(mock_sqs_grating_calibration_run(np.s_[10:20]))
+
+    assert np.isclose(grating_calibration.e0, 990.0, atol=1e-2, rtol=1e-2)
+    assert np.isclose(grating_calibration.slope, 20.0/1000.0, atol=1e-2, rtol=1e-2)
+
+
 @pytest.mark.skipif(not os.path.isdir("/gpfs/exfel/d"), reason="GPFS not available")
 @pytest.mark.vcr
 def test_full_grating2d_calibration_from_data():